requirements.txt
flask
openai
flask_cors
gunicorn
pillow
twilio
flasgger
bcrypt
pymongo
python-dotenv
cryptography
marshmallow